Move field selection cell types into higher types file
